The C program I created and will share
For now, It contains a prototype of my
port of the python 3d renderer code.

it follows roughly the same logic.

Though, it's plagued by memory leaks and the like. Plus the code readability and naming convention is kinda lacking.

Still, pretty proud of this 5 day project, it's a good start and I learned a lot making it. Had fun too.